+++ src/framework/visualization/error.ts
@@ -1,0 +1,6 @@
+import { IntegrationError } from '../execution/error';
+
+export class IntegrationMissingCollectJSON extends Error
+  implements IntegrationError {
+  code = 'MissingCollectJSON';
+}